==> notebooks/01_Fetch_Datasets.ipynb <==
import os

os.chdir("..")

%run -i "hmo_identifier/data/reference.py" "Camden"
%run -i "hmo_identifier/data/open.py" "Camden"
%run -i "hmo_identifier/data/local.py" "Camden"
%run -i "hmo_identifier/data/gla.py" "Camden"
